requirements.txt
kafka-python==2.0.2
pip~=21.0.1
setuptools~=53.0.0
argparse~=1.4.0
cryptography~=3.3.1
lxml~=4.6.2
docutils~=0.16
future~=0.18.2
oauthlib==3.1.0
future~=0.18.2
typed-config==0.1.3
requests-oauthlib==1.3.0
